+++ 42-cursus/ft_printf/ft_printf.h
@@ -21,7 +21,7 @@ int			ft_printf(const char *string, ...);
 void		ft_putcharacter_length(char character, int *length);
 void		ft_hexadecimal(unsigned int x, int *length, char x_or_x);
 void		ft_print_number(int number, int *length);
-void		ft_pointer(size_t pointer, int *length);
+void		ft_pointer(unsigned long pointer, int *length);
 void		ft_string(char *args, int *length);
 void		ft_unsigned_int(unsigned int u, int *length);
 #endif
